+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.27.2] - 2025-05-14
+
 ### Added
 
 - Add `affect_t!` option to `Timestepping.init`.